:construction: NEW: switch to sveltekit / pnpm
