+++ package.json
@@ -1,34 +1,40 @@
 {
-  "name": "react-vite-express-starter",
-  "private": true,
-  "version": "0.0.0",
-  "scripts": {
-    "test": "vitest --watch",
-    "dev": "vite",
-    "dev-node": "tsx watch --inspect src/server",
-    "start": "node dist/server/",
-    "build": "tsc && vite build && tsc -p tsconfig.server.json",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "express": "^4.18.2",
-    "graphql": "^16.6.0",
-    "graphql-yoga": "^3.9.1",
-    "pg": "^8.11.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "remult": "^0.20.3-exp.0"
-  },
-  "devDependencies": {
-    "@graphql-tools/executor-http": "^1.0.0",
-    "@graphql-tools/schema": "^10.0.0",
-    "@types/express": "^4.17.17",
-    "@types/react": "^18.0.27",
-    "@types/react-dom": "^18.0.10",
-    "@vitejs/plugin-react": "^3.1.0",
-    "tsx": "^3.12.3",
-    "typescript": "^4.9.3",
-    "vite": "^4.1.0",
-    "vitest": "^0.31.1"
-  }
+	"name": "graphql-discussion-with-jycouet",
+	"version": "0.0.1",
+	"private": true,
+	"scripts": {
+		"dev": "vite dev --open",
+		"build": "vite build",
+		"preview": "vite preview",
+		"test": "vitest --watch",
+		"test:e2e": "playwright test",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "prettier --plugin-search-dir . --check . && eslint .",
+		"format": "prettier --plugin-search-dir . --write ."
+	},
+	"devDependencies": {
+		"@playwright/test": "^1.28.1",
+		"@sveltejs/adapter-auto": "^2.0.0",
+		"@sveltejs/kit": "^1.5.0",
+		"@typescript-eslint/eslint-plugin": "^5.45.0",
+		"@typescript-eslint/parser": "^5.45.0",
+		"eslint": "^8.28.0",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-plugin-svelte": "^2.26.0",
+		"prettier": "^2.8.0",
+		"prettier-plugin-svelte": "^2.8.1",
+		"svelte": "^3.54.0",
+		"svelte-check": "^3.0.1",
+		"tslib": "^2.4.1",
+		"typescript": "^5.0.0",
+		"vite": "^4.3.0",
+		"vitest": "^0.25.3"
+	},
+	"type": "module",
+	"dependencies": {
+		"express": "^4.18.2",
+		"graphql-yoga": "^3.9.1",
+		"remult": "0.20.3-exp.0"
+	}
 }
