--- src/server/index.ts
@@ -1,25 +1,0 @@
-import express from "express";
-import { api } from "./api";
-
-// import { remultGraphql } from "remult/graphql";
-import { remultGraphql } from "./graphql";
-import { createSchema, createYoga } from "graphql-yoga";
-
-const app = express();
-app.use(api);
-const { schema, resolvers } = remultGraphql(api);
-
-const yoga = createYoga({
-  graphiql: true,
-  graphqlEndpoint: "/api/yoga",
-  schema: createSchema({
-    typeDefs: schema,
-    resolvers,
-  }),
-});
-// Bind GraphQL Yoga to the graphql endpoint to avoid rendering the playground on any path
-app.use(yoga.graphqlEndpoint, yoga);
-
-app.listen(3002, () =>
-  console.log("Server started on http://localhost:3002/api/yoga")
-);
