+++ src/shared/Category.ts
@@ -1,11 +1,9 @@
-import { Entity, Fields } from "remult";
+import { Entity, Fields } from 'remult';
 
-@Entity("categories", {
-  allowApiCrud: true,
-})
+@Entity('categories', { allowApiCrud: true })
 export class Category {
-  @Fields.cuid()
-  id = "";
-  @Fields.string()
-  name = "";
+	@Fields.cuid()
+	id = '';
+	@Fields.string()
+	name = '';
 }
