--- tsconfig.server.json
@@ -1,14 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "module": "commonjs",
-        "emitDecoratorMetadata": true,
-        "esModuleInterop": true,
-        "noEmit": false,
-        "outDir": "dist",
-        "rootDir": "src"
-    },
-    "include": [
-        "src/server/index.ts"
-    ]
-}
